+++ protected/humhub/modules/tour/widgets/views/dashboard.php
@@ -6,7 +6,6 @@ use humhub\modules\content\components\ContentContainerSettingsManager;
 use humhub\modules\tour\assets\TourAsset;
 use humhub\modules\tour\TourConfig;
 use humhub\modules\ui\icon\widgets\Icon;
-use humhub\widgets\bootstrap\Link;
 use humhub\widgets\PanelMenu;
 use yii\helpers\Url;
 
@@ -18,32 +17,14 @@ TourAsset::register($this);
  * @var bool $showWelcome
  */
 
-$removeOptionHtml = Html::tag(
-    'li',
-    Link::to(
-        Yii::t('TourModule.base', '<strong>Remove</strong> tour panel'),
-        ["/tour/tour/hide-panel", "ajax" => 1],
-    )
-        ->encodeLabel(false)
-        ->icon('eye-slash')
-        ->action('tour.hidePanel')
-        ->confirm(
-            Icon::get('eye-slash') . ' ' . Yii::t('TourModule.base', ' Remove panel'),
-            Yii::t('TourModule.base', 'This action will remove the tour panel from your dashboard. You can reactivate it at<br>Account settings <i class="fa fa-caret-right"></i> Settings.'),
-            Yii::t('TourModule.base', 'Ok'),
-            Yii::t('TourModule.base', 'Cancel'),
-        )
-        ->cssClass(['dropdown-item']),
-);
+$title = Yii::t('TourModule.base', '<strong>Getting</strong> Started');
 ?>
 
 <div class="panel panel-default panel-tour" id="getting-started-panel">
-    <?= PanelMenu::widget([
-        'extraMenus' => $removeOptionHtml,
-    ]) ?>
+    <?= PanelMenu::widget(['panelLabel' => $title]) ?>
 
     <div class="panel-heading">
-        <?= Yii::t('TourModule.base', '<strong>Getting</strong> Started') ?>
+        <?= $title ?>
     </div>
 
     <div class="panel-body">