+++ CHANGELOG.md
@@ -16,6 +16,7 @@ HumHub Changelog
 - Enh #5141: Added CLI command to delete disabled or soft deleted users
 - Enh #8100: Move `static` and `themes` into `protected/humhub`
 - Enh #8128: Exact search for user id in admin user search
+- Enh #8138: Right panel widgets: add a "Hide" menu item
 
 1.18.3 (TBD)
 ------------